6_network_programming/requirements.txt
speedtest-cli==2.1.3
getmac==0.9.4